defer: Move duplicate inlineCallbacks tracebacks test to correct module
